Fix documentList
