--- src/components/ListItem.js
@@ -1,35 +1,0 @@
-import React, { Component } from 'react';
-import { Text, TouchableWithoutFeedback, View } from 'react-native';
-import { Actions } from 'react-native-router-flux';
-import { CardSection } from './common';
-
-class ListItem extends Component {
-  onRowPress() {
-    Actions.documentEdit({ document: this.props.document });
-  }
-
-  render() {
-    const { name } = this.props.document;
-
-    return (
-      <TouchableWithoutFeedback onPress={this.onRowPress.bind(this)}>
-        <View>
-          <CardSection>
-            <Text style={styles.titleStyle}>
-              {name}
-            </Text>
-          </CardSection>
-        </View>
-      </TouchableWithoutFeedback>
-    );
-  }
-}
-
-const styles = {
-  titleStyle: {
-    fontSize: 18,
-    paddingLeft: 15
-  }
-};
-
-export default ListItem;
